requirements.txt
keras==2.12.0
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.2
python_dateutil==2.8.2
scikit_learn==1.2.2
streamlit==1.23.1
yfinance==0.2.18
tensorflow==2.12.0
